requirements.txt
colorama==0.4.1
Flask==1.0.2
influxdb==5.2.2
transitions==0.6.9
requests==2.22.0
urllib3[secure]==1.25.3
pint==0.9
psutil==5.6.5
python-pushover==0.4
pyvisa==1.9.1
pyyaml==5.1.1
pyzmq==18.1.0
wrapt==1.11.2
